Ran destiny, reorganized /src folder
