+++ src/bhotm/shared/bhotmEntryContent.js
